api/src/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.1
Flask==1.1.2
Flask-Hashing==1.1
Flask-JWT-Extended==3.24.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
PyJWT==1.7.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==0.16.0
zipp==3.1.0
